+++ azurectl/commands/storage_container.py
@@ -12,62 +12,35 @@
 # limitations under the License.
 #
 """
-usage: azurectl compute storage -h | --help
-       azurectl compute storage container list
-       azurectl compute storage container show
+usage: azurectl storage container -h | --help
+       azurectl storage container list
+       azurectl storage container show
            [--name=<containername>]
-       azurectl compute storage container create
+       azurectl storage container create
            [--name=<containername>]
-       azurectl compute storage container delete --name=<containername>
-       azurectl compute storage container sas
+       azurectl storage container delete --name=<containername>
+       azurectl storage container sas
            [--name=<containername>]
            [--start-datetime=<start>]
            [--expiry-datetime=<expiry>]
            [--permissions=<permissions>]
-       azurectl compute storage share list
-       azurectl compute storage share create --name=<sharename>
-       azurectl compute storage share delete --name=<sharename>
-       azurectl compute storage upload --source=<file>
-           [--name=<blobname>]
-           [--max-chunk-size=<size>]
-           [--quiet]
-       azurectl compute storage delete --name=<blobname>
-       azurectl compute storage container help
-       azurectl compute storage share help
-       azurectl compute storage help
+       azurectl storage container help
 
 commands:
-    container list
-        list storage container names for configured account
-    container show
+    list
+        list container names for configured account
+    show
         show container content for configured account and container
-    container create
-        create container with the specified name
-    container delete
-        delete container with the specified name
-    upload
-        upload xz compressed blob to the given container
-    delete
-        storage: delete blob from the given container
-        storage share: delete share from the storage account
     create
-        create file share in the storage account
-    container help
-        show manual page for container sub command
-    share help
-        show manual page for share sub command
+        create container with the specified name
+    delete
+        delete container with the specified name
     help
-        show manual page for storage command
+        show manual page for container command
 
 options:
-    --source=<file>
-        file to upload
     --name=<name>
-        blobname: name of the file in the storage pool
-        sharename: name of the files share
-        containername: name of the container
-    --max-chunk-size=<size>
-        max chunk size in bytes for upload, default 4MB
+        name of the container
     --start-datetime=<start>
         Date (and optionally time) to grant access via a shared access
         signature. [default: now]
@@ -82,14 +55,10 @@ options:
         d  Delete
         l  List
         [default: rl]
-    --quiet
-        suppress progress information on upload
 """
 import datetime
 import dateutil.parser
 import re
-from apscheduler.schedulers.background import BackgroundScheduler
-from pytz import utc
 
 # project
 from base import CliTask
@@ -97,16 +66,17 @@ from ..account.service import AzureAccount
 from ..utils.collector import DataCollector
 from ..utils.output import DataOutput
 from ..logger import log
-from ..azurectl_exceptions import AzureInvalidCommand
-from ..storage.storage import Storage
 from ..storage.container import Container
-from ..storage.fileshare import FileShare
 from ..help import Help
 
+from ..azurectl_exceptions import (
+    AzureInvalidCommand
+)
 
-class ComputeStorageTask(CliTask):
+
+class StorageContainerTask(CliTask):
     """
-        Process storage commands
+        Process container commands
     """
     def process(self):
         self.manual = Help()
@@ -140,37 +110,23 @@ class ComputeStorageTask(CliTask):
 
         self.__validate_permissions_arg()
 
-        self.storage = Storage(self.account, container_name)
         self.container = Container(self.account)
-        self.fileshare = FileShare(self.account)
 
-        if self.command_args['container'] and self.command_args['list']:
+        if self.command_args['list']:
             self.__container_list()
-        elif self.command_args['container'] and self.command_args['show']:
+        elif self.command_args['show']:
             self.__container_content(container_name)
-        elif self.command_args['container'] and self.command_args['create']:
+        elif self.command_args['create']:
             self.__container_create(container_name)
-        elif self.command_args['container'] and self.command_args['delete']:
+        elif self.command_args['delete']:
             self.__container_delete(container_name)
-        elif self.command_args['container'] and self.command_args['sas']:
+        elif self.command_args['sas']:
             self.__container_sas(
                 container_name,
                 start,
                 expiry,
                 self.command_args['--permissions']
             )
-        elif self.command_args['share'] and self.command_args['list']:
-            self.__share_list()
-        elif self.command_args['share'] and self.command_args['create']:
-            self.__share_create()
-        elif self.command_args['share'] and self.command_args['delete']:
-            self.__share_delete()
-        elif self.command_args['upload']:
-            self.__upload()
-        elif self.command_args['delete']:
-            self.__delete()
-
-    # argument validation
 
     def __validate_date_arg(self, cmd_arg):
         try:
@@ -187,59 +143,12 @@ class ComputeStorageTask(CliTask):
                 '--permissions=' + self.command_args['--permissions']
             )
 
-    # tasks
-
     def __help(self):
-        if self.command_args['container'] and self.command_args['help']:
-            self.manual.show('azurectl::compute::storage::container')
-        elif self.command_args['share'] and self.command_args['help']:
-            self.manual.show('azurectl::compute::storage::share')
-        elif self.command_args['help']:
-            self.manual.show('azurectl::compute::storage')
+        if self.command_args['help']:
+            self.manual.show('azurectl::storage::container')
         else:
             return False
         return self.manual
-
-    def __upload(self):
-        if self.command_args['--quiet']:
-            self.__upload_no_progress()
-        else:
-            self.__upload_with_progress()
-
-    def __upload_no_progress(self):
-        try:
-            self.__process_upload()
-        except (KeyboardInterrupt):
-            raise SystemExit('azurectl aborted by keyboard interrupt')
-
-    def __upload_with_progress(self):
-        image = self.command_args['--source']
-        progress = BackgroundScheduler(timezone=utc)
-        progress.add_job(
-            self.storage.print_upload_status, 'interval', seconds=3
-        )
-        progress.start()
-        try:
-            self.__process_upload()
-            self.storage.print_upload_status()
-            progress.shutdown()
-        except (KeyboardInterrupt):
-            progress.shutdown()
-            raise SystemExit('azurectl aborted by keyboard interrupt')
-        print
-        log.info('Uploaded %s', image)
-
-    def __process_upload(self):
-        self.storage.upload(
-            self.command_args['--source'],
-            self.command_args['--name'],
-            self.command_args['--max-chunk-size']
-        )
-
-    def __delete(self):
-        image = self.command_args['--name']
-        self.storage.delete(image)
-        log.info('Deleted %s', image)
 
     def __container_sas(self, container_name, start, expiry, permissions):
         if self.command_args['--name']:
@@ -279,17 +188,3 @@ class ComputeStorageTask(CliTask):
         log.info('Request to delete container %s', container_name)
         self.container.delete(container_name)
         log.info('Deleted %s container', container_name)
-
-    def __share_list(self):
-        self.result.add('share_names', self.fileshare.list())
-        self.out.display()
-
-    def __share_create(self):
-        share_name = self.command_args['--name']
-        self.fileshare.create(share_name)
-        log.info('Created Files Share %s', share_name)
-
-    def __share_delete(self):
-        share_name = self.command_args['--name']
-        self.fileshare.delete(share_name)
-        log.info('Deleted Files Share %s', share_name)
