+++ test/unit/cli_test.py
@@ -11,12 +11,12 @@ class TestCli:
     def setup(self):
         self.help_global_args = {
             '--output-format': None,
-            'compute': True,
             'help': False,
             '--region': None,
             '--storage-container': None,
+            'compute': False,
             'setup': False,
-            'storage': False,
+            'storage': True,
             '-h': False,
             '--output-style': None,
             '--config': None,
@@ -29,27 +29,21 @@ class TestCli:
         self.help_command_args = {
             '--expiry-datetime': '30 days from start',
             '--help': False,
-            '--max-chunk-size': None,
             '--name': None,
             '--permissions': 'rl',
-            '--quiet': False,
-            '--source': None,
             '--start-datetime': 'now',
             '-h': False,
-            'compute': True,
             'container': True,
             'create': False,
             'delete': False,
             'help': False,
             'list': True,
             'sas': False,
-            'share': False,
             'show': False,
-            'storage': True,
-            'upload': False
+            'storage': True
         }
         sys.argv = [
-            sys.argv[0], 'compute', 'storage', 'container', 'list'
+            sys.argv[0], 'storage', 'container', 'list'
         ]
         self.cli = Cli()
         self.loaded_command = self.cli.load_command()
@@ -58,10 +52,10 @@ class TestCli:
         assert self.cli.show_help() is False
 
     def test_get_servicename(self):
-        assert self.cli.get_servicename() == 'compute'
+        assert self.cli.get_servicename() == 'storage'
 
     def test_get_command(self):
-        assert self.cli.get_command() == 'storage'
+        assert self.cli.get_command() == 'container'
 
     def test_get_command_args(self):
         assert self.cli.get_command_args() == self.help_command_args
@@ -93,4 +87,5 @@ class TestCli:
     def test_get_servicename_unknown(self):
         self.cli.all_args['compute'] = False
         self.cli.all_args['setup'] = False
+        self.cli.all_args['storage'] = False
         self.cli.get_servicename()
